Proactively change to float32 in metadata helper apply_raster_scale_and_offset_to_band_metadata
Still to be fixed in openeo-geotrellis-extensions, but it doesn't hurt
to already update this metadata field
(only used on experimental collection metadata at the moment)

refs https://github.com/Open-EO/openeo-geotrellis-extensions/issues/761
https://github.com/Open-EO/openeo-geopyspark-driver/issues/1666
https://github.com/eu-cdse/openeo-cdse-infra/issues/686

+++ openeogeotrellis/catalog/manage.py
@@ -503,7 +503,7 @@ def apply_raster_scale_and_offset_to_band_metadata(bands: List[BandMetadata]) ->
         data["raster_scale"] = None
         data["raster_offset"] = None
         if to_float:
-            data["data_type"] = "float64"
+            data["data_type"] = "float32"
             # TODO: possible to set `nodata`? e.g. "nan"?
             data["nodata"] = None
